requirements.txt
Django==4.2.1
djangorestframework==3.14.0
requests==2.30.0
setuptools==23.0.1
wheel
twine
django-rest-swagger==2.2.0
black
factory_boy
pytest-django
pytest-factoryboy